[nextest-runner] move partition tests into their own file
A bit hard to read with `test_list.rs`.
