+++ nextest-runner/src/list/test_list.rs
@@ -348,7 +348,7 @@ impl<'g> TestList<'g> {
     /// Creates a new test list with the given binary names and outputs.
     #[cfg(test)]
     #[expect(clippy::too_many_arguments)]
-    fn new_with_outputs(
+    pub(crate) fn new_with_outputs(
         test_bin_outputs: impl IntoIterator<
             Item = (RustTestArtifact<'g>, impl AsRef<str>, impl AsRef<str>),
         >,
@@ -1756,18 +1756,20 @@ mod tests {
     use crate::{
         cargo_config::{TargetDefinitionLocation, TargetTriple, TargetTripleSource},
         config::scripts::{ScriptCommand, ScriptCommandRelativeTo},
-        list::SerializableFormat,
+        list::{
+            SerializableFormat,
+            test_helpers::{PACKAGE_GRAPH_FIXTURE, package_metadata},
+        },
         platform::{BuildPlatforms, HostPlatform, PlatformLibdir, TargetPlatform},
         target_runner::PlatformRunnerSource,
         test_filter::{RunIgnored, TestFilterPatterns},
     };
-    use guppy::CargoMetadata;
     use iddqd::id_ord_map;
     use indoc::indoc;
     use nextest_filtering::{CompiledExpr, Filterset, FiltersetKind, ParseContext};
     use nextest_metadata::{FilterMatch, MismatchReason, PlatformLibdirUnavailable, RustTestKind};
     use pretty_assertions::assert_eq;
-    use std::{hash::DefaultHasher, sync::LazyLock};
+    use std::hash::DefaultHasher;
     use target_spec::Platform;
     use test_strategy::proptest;
 
@@ -2142,567 +2144,6 @@ mod tests {
         );
     }
 
-    /// Verifies that per-binary partitioning (applied as post-processing)
-    /// produces the same results as the old inline approach: count:1/2 should
-    /// select every other test, independently within each binary.
-    #[test]
-    fn test_apply_per_binary_partitioning_count() {
-        let binary1_output = indoc::indoc! {"
-            alpha: test
-            beta: test
-            gamma: test
-            delta: test
-        "};
-        let binary2_output = indoc::indoc! {"
-            one: test
-            two: test
-            three: test
-        "};
-
-        let test_filter = TestFilter::new(
-            NextestRunMode::Test,
-            RunIgnored::Default,
-            TestFilterPatterns::default(),
-            Vec::new(),
-        )
-        .unwrap();
-
-        let partitioner = PartitionerBuilder::Count {
-            shard: 1,
-            total_shards: 2,
-        };
-
-        let ecx = simple_ecx();
-        let test_list = TestList::new_with_outputs(
-            [
-                (make_test_artifact("pkg::binary1"), binary1_output, ""),
-                (make_test_artifact("pkg::binary2"), binary2_output, ""),
-            ],
-            Utf8PathBuf::from("/fake/path"),
-            simple_build_meta(),
-            &test_filter,
-            Some(&partitioner),
-            EnvironmentMap::empty(),
-            &ecx,
-            FilterBound::All,
-        )
-        .expect("valid output");
-
-        // count:1/2 selects every other test (index 0, 2, ...) within each
-        // binary independently.
-        //
-        // Binary 1 (sorted: alpha, beta, delta, gamma): shard 1 gets alpha, delta.
-        // Binary 2 (sorted: one, three, two): shard 1 gets one, two.
-        let binary1_suite = test_list
-            .rust_suites
-            .get(&RustBinaryId::new("pkg::binary1"))
-            .expect("binary1 should exist");
-        assert_eq!(
-            collect_matching_tests(binary1_suite),
-            vec!["alpha", "delta"]
-        );
-
-        let binary2_suite = test_list
-            .rust_suites
-            .get(&RustBinaryId::new("pkg::binary2"))
-            .expect("binary2 should exist");
-        assert_eq!(collect_matching_tests(binary2_suite), vec!["one", "two"]);
-    }
-
-    /// Verifies that non-ignored and ignored tests are partitioned
-    /// independently: adding or removing an ignored test does not change which
-    /// non-ignored tests are selected, and vice versa.
-    #[test]
-    fn test_partition_ignored_independent() {
-        // 4 non-ignored tests, 3 ignored tests.
-        let non_ignored_output = indoc::indoc! {"
-            alpha: test
-            beta: test
-            gamma: test
-            delta: test
-        "};
-        let ignored_output = indoc::indoc! {"
-            ig_one: test
-            ig_two: test
-            ig_three: test
-        "};
-
-        // Use RunIgnored::All so both non-ignored and ignored tests are included.
-        let test_filter = TestFilter::new(
-            NextestRunMode::Test,
-            RunIgnored::All,
-            TestFilterPatterns::default(),
-            Vec::new(),
-        )
-        .unwrap();
-
-        let partitioner = PartitionerBuilder::Count {
-            shard: 1,
-            total_shards: 2,
-        };
-
-        let ecx = simple_ecx();
-        let test_list = TestList::new_with_outputs(
-            [(
-                make_test_artifact("pkg::binary"),
-                non_ignored_output,
-                ignored_output,
-            )],
-            Utf8PathBuf::from("/fake/path"),
-            simple_build_meta(),
-            &test_filter,
-            Some(&partitioner),
-            EnvironmentMap::empty(),
-            &ecx,
-            FilterBound::All,
-        )
-        .expect("valid output");
-
-        let suite = test_list
-            .rust_suites
-            .get(&RustBinaryId::new("pkg::binary"))
-            .expect("binary should exist");
-
-        // Non-ignored (sorted: alpha, beta, delta, gamma): shard 1 picks
-        // indices 0, 2 => alpha, delta.
-        let non_ignored_matches: Vec<_> = suite
-            .status
-            .test_cases()
-            .filter(|tc| !tc.test_info.ignored && tc.test_info.filter_match.is_match())
-            .map(|tc| tc.name.as_str())
-            .collect();
-        assert_eq!(non_ignored_matches, vec!["alpha", "delta"]);
-
-        // Ignored (sorted: ig_one, ig_three, ig_two): shard 1 picks indices
-        // 0, 2 => ig_one, ig_two. Independent of the non-ignored partitioner.
-        let ignored_matches: Vec<_> = suite
-            .status
-            .test_cases()
-            .filter(|tc| tc.test_info.ignored && tc.test_info.filter_match.is_match())
-            .map(|tc| tc.name.as_str())
-            .collect();
-        assert_eq!(ignored_matches, vec!["ig_one", "ig_two"]);
-    }
-
-    /// Verifies that `RerunAlreadyPassed` tests participate in partition
-    /// counting (to maintain stable shard assignment) but their status is
-    /// preserved as `RerunAlreadyPassed`, not changed to `Partition`.
-    #[test]
-    fn test_partition_rerun_already_passed() {
-        use crate::record::{ComputedRerunInfo, RerunTestSuiteInfo};
-
-        let binary_id = RustBinaryId::new("pkg::binary");
-        let non_ignored_output = indoc::indoc! {"
-            alpha: test
-            beta: test
-            gamma: test
-            delta: test
-        "};
-        let partitioner = PartitionerBuilder::Count {
-            shard: 1,
-            total_shards: 2,
-        };
-        let ecx = simple_ecx();
-
-        // Mark "beta" as already passed in a prior rerun. "beta" is at sorted
-        // index 1. The filter will mark it RerunAlreadyPassed before
-        // partitioning runs.
-        let rerun_suite = RerunTestSuiteInfo {
-            binary_id: binary_id.clone(),
-            passing: BTreeSet::from([TestCaseName::new("beta")]),
-            outstanding: BTreeSet::from([
-                TestCaseName::new("alpha"),
-                TestCaseName::new("delta"),
-                TestCaseName::new("gamma"),
-            ]),
-        };
-        let mut rerun_filter = TestFilter::new(
-            NextestRunMode::Test,
-            RunIgnored::Default,
-            TestFilterPatterns::default(),
-            Vec::new(),
-        )
-        .unwrap();
-        rerun_filter.set_outstanding_tests(ComputedRerunInfo {
-            test_suites: iddqd::id_ord_map! { rerun_suite },
-        });
-
-        let rerun_list = TestList::new_with_outputs(
-            [(make_test_artifact("pkg::binary"), non_ignored_output, "")],
-            Utf8PathBuf::from("/fake/path"),
-            simple_build_meta(),
-            &rerun_filter,
-            Some(&partitioner),
-            EnvironmentMap::empty(),
-            &ecx,
-            FilterBound::All,
-        )
-        .expect("valid output");
-
-        let suite = rerun_list
-            .rust_suites
-            .get(&binary_id)
-            .expect("binary should exist");
-
-        // After filtering, test states are (sorted: alpha, beta, delta, gamma):
-        //   alpha (index 0): Matches
-        //   beta  (index 1): RerunAlreadyPassed
-        //   delta (index 2): Matches
-        //   gamma (index 3): Matches
-        //
-        // The count:1/2 partitioner sees 4 tests (including beta, which
-        // participates in counting). Shard 1 picks indices 0 and 2:
-        //   alpha (0): Matches -> kept by partitioner
-        //   beta  (1): RerunAlreadyPassed -> counted, status preserved
-        //   delta (2): Matches -> kept by partitioner
-        //   gamma (3): Matches -> excluded by partitioner (Partition mismatch)
-        let rerun_results: Vec<_> = suite
-            .status
-            .test_cases()
-            .map(|tc| (tc.name.as_str(), tc.test_info.filter_match))
-            .collect();
-
-        assert_eq!(
-            rerun_results,
-            vec![
-                ("alpha", FilterMatch::Matches),
-                (
-                    "beta",
-                    FilterMatch::Mismatch {
-                        reason: MismatchReason::RerunAlreadyPassed,
-                    }
-                ),
-                ("delta", FilterMatch::Matches),
-                (
-                    "gamma",
-                    FilterMatch::Mismatch {
-                        reason: MismatchReason::Partition,
-                    }
-                ),
-            ]
-        );
-
-        // If beta were not counted by the partitioner, the partitioner would
-        // only see 3 Matches tests (alpha, delta, gamma) and shard 1 would
-        // pick alpha(0) and gamma(2) instead of alpha(0) and delta(2). Build
-        // the counterfactual to confirm the results actually differ.
-        let no_rerun_filter = TestFilter::new(
-            NextestRunMode::Test,
-            RunIgnored::Default,
-            TestFilterPatterns::default(),
-            Vec::new(),
-        )
-        .unwrap();
-
-        let no_rerun_list = TestList::new_with_outputs(
-            [(make_test_artifact("pkg::binary"), non_ignored_output, "")],
-            Utf8PathBuf::from("/fake/path"),
-            simple_build_meta(),
-            &no_rerun_filter,
-            Some(&partitioner),
-            EnvironmentMap::empty(),
-            &ecx,
-            FilterBound::All,
-        )
-        .expect("valid output");
-
-        let suite = no_rerun_list
-            .rust_suites
-            .get(&binary_id)
-            .expect("binary should exist");
-
-        // Without rerun info, all 4 tests are Matches. The partitioner sees
-        // all 4, and shard 1 still picks indexes 0 and 2: alpha and delta.
-        // This is the same result as the rerun case, confirming that the
-        // RerunAlreadyPassed test (beta) participated in counting and kept
-        // shard assignments stable.
-        let no_rerun_results: Vec<_> = suite
-            .status
-            .test_cases()
-            .map(|tc| (tc.name.as_str(), tc.test_info.filter_match))
-            .collect();
-
-        assert_eq!(
-            no_rerun_results,
-            vec![
-                ("alpha", FilterMatch::Matches),
-                (
-                    "beta",
-                    FilterMatch::Mismatch {
-                        reason: MismatchReason::Partition,
-                    }
-                ),
-                ("delta", FilterMatch::Matches),
-                (
-                    "gamma",
-                    FilterMatch::Mismatch {
-                        reason: MismatchReason::Partition,
-                    }
-                ),
-            ]
-        );
-
-        // The key property: the set of Matches tests that actually run is
-        // {alpha, delta} in both cases. The RerunAlreadyPassed counting
-        // ensures shard stability across reruns.
-        let rerun_running: Vec<_> = rerun_results
-            .iter()
-            .filter(|(_, fm)| fm.is_match())
-            .map(|(name, _)| *name)
-            .collect();
-        let no_rerun_running: Vec<_> = no_rerun_results
-            .iter()
-            .filter(|(_, fm)| fm.is_match())
-            .map(|(name, _)| *name)
-            .collect();
-        assert_eq!(rerun_running, no_rerun_running);
-    }
-
-    /// Verifies that tests filtered out by name patterns do not participate
-    /// in partition counting. If pre-filtered tests were counted, shard
-    /// assignments would shift when name filters change.
-    #[test]
-    fn test_partition_prefiltered_excluded_from_counting() {
-        let non_ignored_output = indoc::indoc! {"
-            alpha: test
-            beta: test
-            gamma: test
-            delta: test
-            epsilon: test
-        "};
-
-        // Filter to only tests containing "a" (alpha, beta, gamma, delta match;
-        // epsilon does not).
-        let test_filter = TestFilter::new(
-            NextestRunMode::Test,
-            RunIgnored::Default,
-            TestFilterPatterns::new(vec!["a".to_string()]),
-            Vec::new(),
-        )
-        .unwrap();
-
-        let partitioner = PartitionerBuilder::Count {
-            shard: 1,
-            total_shards: 2,
-        };
-
-        let ecx = simple_ecx();
-        let filtered_list = TestList::new_with_outputs(
-            [(make_test_artifact("pkg::binary"), non_ignored_output, "")],
-            Utf8PathBuf::from("/fake/path"),
-            simple_build_meta(),
-            &test_filter,
-            Some(&partitioner),
-            EnvironmentMap::empty(),
-            &ecx,
-            FilterBound::All,
-        )
-        .expect("valid output");
-
-        let suite = filtered_list
-            .rust_suites
-            .get(&RustBinaryId::new("pkg::binary"))
-            .expect("binary should exist");
-
-        // Pre-filter: sorted tests are alpha, beta, delta, epsilon, gamma.
-        // After name filter, epsilon is Mismatch(String). The remaining 4
-        // matching tests (alpha, beta, delta, gamma) are partitioned by
-        // count:1/2, which picks indices 0 and 2 => alpha, delta.
-        assert_eq!(collect_matching_tests(suite), vec!["alpha", "delta"]);
-
-        // Verify epsilon was filtered out by name, not by partition.
-        let epsilon = suite
-            .status
-            .get(&TestCaseName::new("epsilon"))
-            .expect("epsilon should exist");
-        assert_eq!(
-            epsilon.test_info.filter_match,
-            FilterMatch::Mismatch {
-                reason: MismatchReason::String,
-            }
-        );
-
-        // Now build a list without name filters. All 5 tests are Matches, and
-        // count:1/2 picks indices 0, 2, 4 => alpha, delta, gamma.
-        let no_name_filter = TestFilter::new(
-            NextestRunMode::Test,
-            RunIgnored::Default,
-            TestFilterPatterns::default(),
-            Vec::new(),
-        )
-        .unwrap();
-
-        let unfiltered_list = TestList::new_with_outputs(
-            [(make_test_artifact("pkg::binary"), non_ignored_output, "")],
-            Utf8PathBuf::from("/fake/path"),
-            simple_build_meta(),
-            &no_name_filter,
-            Some(&partitioner),
-            EnvironmentMap::empty(),
-            &ecx,
-            FilterBound::All,
-        )
-        .expect("valid output");
-
-        let suite = unfiltered_list
-            .rust_suites
-            .get(&RustBinaryId::new("pkg::binary"))
-            .expect("binary should exist");
-
-        // With 5 tests and no name filter, count:1/2 picks indices 0, 2, 4 =>
-        // alpha, delta, gamma. This differs from the filtered case (alpha,
-        // delta), proving that the pre-filtered epsilon did not participate in
-        // partition counting.
-        assert_eq!(
-            collect_matching_tests(suite),
-            vec!["alpha", "delta", "gamma"]
-        );
-    }
-
-    /// Verifies that hash-based partitioning works correctly when wired
-    /// through `TestList::new_with_outputs`. Since hash partitioning is
-    /// stateless, the results should be deterministic and independent of
-    /// binary boundaries.
-    #[test]
-    fn test_partition_hash() {
-        let non_ignored_output = indoc::indoc! {"
-            alpha: test
-            beta: test
-            gamma: test
-            delta: test
-        "};
-
-        let test_filter = TestFilter::new(
-            NextestRunMode::Test,
-            RunIgnored::Default,
-            TestFilterPatterns::default(),
-            Vec::new(),
-        )
-        .unwrap();
-
-        // hash:1/2 and hash:2/2 should partition all tests between them.
-        let partitioner_shard1 = PartitionerBuilder::Hash {
-            shard: 1,
-            total_shards: 2,
-        };
-        let partitioner_shard2 = PartitionerBuilder::Hash {
-            shard: 2,
-            total_shards: 2,
-        };
-
-        let ecx = simple_ecx();
-        let list_shard1 = TestList::new_with_outputs(
-            [(make_test_artifact("pkg::binary"), non_ignored_output, "")],
-            Utf8PathBuf::from("/fake/path"),
-            simple_build_meta(),
-            &test_filter,
-            Some(&partitioner_shard1),
-            EnvironmentMap::empty(),
-            &ecx,
-            FilterBound::All,
-        )
-        .expect("valid output");
-
-        let list_shard2 = TestList::new_with_outputs(
-            [(make_test_artifact("pkg::binary"), non_ignored_output, "")],
-            Utf8PathBuf::from("/fake/path"),
-            simple_build_meta(),
-            &test_filter,
-            Some(&partitioner_shard2),
-            EnvironmentMap::empty(),
-            &ecx,
-            FilterBound::All,
-        )
-        .expect("valid output");
-
-        let suite1 = list_shard1
-            .rust_suites
-            .get(&RustBinaryId::new("pkg::binary"))
-            .expect("binary should exist");
-        let suite2 = list_shard2
-            .rust_suites
-            .get(&RustBinaryId::new("pkg::binary"))
-            .expect("binary should exist");
-
-        let shard1_matches = collect_matching_tests(suite1);
-        let shard2_matches = collect_matching_tests(suite2);
-
-        // The two shards should be disjoint and together cover all tests.
-        let mut all_tests: Vec<&str> = shard1_matches
-            .iter()
-            .chain(shard2_matches.iter())
-            .copied()
-            .collect();
-        all_tests.sort();
-        assert_eq!(all_tests, vec!["alpha", "beta", "delta", "gamma"]);
-
-        // Each shard should have at least one test (with 4 tests and 2 shards,
-        // a proper hash function won't put all tests in one shard).
-        assert!(
-            !shard1_matches.is_empty() && !shard2_matches.is_empty(),
-            "both shards should have tests: shard1={shard1_matches:?}, shard2={shard2_matches:?}"
-        );
-
-        // The shards must not overlap.
-        for test in &shard1_matches {
-            assert!(
-                !shard2_matches.contains(test),
-                "test {test} should not appear in both shards"
-            );
-        }
-    }
-
-    /// Verifies that `test_count` and `run_count` are correct after
-    /// partitioning: `test_count` counts all tests (including partitioned-out
-    /// ones), while `run_count` excludes skipped tests.
-    #[test]
-    fn test_partition_test_count_and_run_count() {
-        let non_ignored_output = indoc::indoc! {"
-            alpha: test
-            beta: test
-            gamma: test
-            delta: test
-        "};
-
-        let test_filter = TestFilter::new(
-            NextestRunMode::Test,
-            RunIgnored::Default,
-            TestFilterPatterns::default(),
-            Vec::new(),
-        )
-        .unwrap();
-
-        let partitioner = PartitionerBuilder::Count {
-            shard: 1,
-            total_shards: 2,
-        };
-
-        let ecx = simple_ecx();
-        let test_list = TestList::new_with_outputs(
-            [(make_test_artifact("pkg::binary"), non_ignored_output, "")],
-            Utf8PathBuf::from("/fake/path"),
-            simple_build_meta(),
-            &test_filter,
-            Some(&partitioner),
-            EnvironmentMap::empty(),
-            &ecx,
-            FilterBound::All,
-        )
-        .expect("valid output");
-
-        // test_count includes all tests (matching and non-matching).
-        assert_eq!(test_list.test_count(), 4);
-
-        // count:1/2 selects 2 out of 4 tests (alpha, delta).
-        assert_eq!(test_list.run_count(), 2);
-
-        // Verify the invariant: run_count + skip_count == test_count.
-        assert_eq!(
-            test_list.run_count() + test_list.skip_counts().skipped_tests,
-            test_list.test_count()
-        );
-    }
-
     #[test]
     fn apply_wrappers_examples() {
         cfg_if::cfg_if! {
@@ -2963,65 +2404,6 @@ mod tests {
                 vec![wrapper_path, "--verbose", "binary", "arg"],
             );
         }
-    }
-
-    /// Creates a test artifact with the given binary ID and sensible defaults.
-    fn make_test_artifact(binary_id: &str) -> RustTestArtifact<'static> {
-        let binary_name = binary_id.rsplit("::").next().unwrap_or(binary_id);
-        RustTestArtifact {
-            binary_path: format!("/fake/{binary_name}").into(),
-            cwd: "/fake/cwd".into(),
-            package: package_metadata(),
-            binary_name: binary_name.to_owned(),
-            binary_id: RustBinaryId::new(binary_id),
-            kind: RustTestBinaryKind::LIB,
-            non_test_binaries: BTreeSet::new(),
-            build_platform: BuildPlatform::Target,
-        }
-    }
-
-    /// Creates a minimal build meta for tests that don't need cross-compilation.
-    fn simple_build_meta() -> RustBuildMeta<TestListState> {
-        let build_platforms = BuildPlatforms {
-            host: HostPlatform {
-                platform: TargetTriple::x86_64_unknown_linux_gnu().platform,
-                libdir: PlatformLibdir::Available("/fake/libdir".into()),
-            },
-            target: None,
-        };
-        RustBuildMeta::new("/fake", build_platforms).map_paths(&PathMapper::noop())
-    }
-
-    /// Creates a default eval context using `CompiledExpr::ALL`.
-    fn simple_ecx() -> EvalContext<'static> {
-        EvalContext {
-            default_filter: &CompiledExpr::ALL,
-        }
-    }
-
-    /// Collects the names of tests that match filters from a suite.
-    fn collect_matching_tests<'a>(suite: &'a RustTestSuite<'_>) -> Vec<&'a str> {
-        suite
-            .status
-            .test_cases()
-            .filter(|tc| tc.test_info.filter_match.is_match())
-            .map(|tc| tc.name.as_str())
-            .collect()
-    }
-
-    static PACKAGE_GRAPH_FIXTURE: LazyLock<PackageGraph> = LazyLock::new(|| {
-        static FIXTURE_JSON: &str = include_str!("../../../fixtures/cargo-metadata.json");
-        let metadata = CargoMetadata::parse_json(FIXTURE_JSON).expect("fixture is valid JSON");
-        metadata
-            .build_graph()
-            .expect("fixture is valid PackageGraph")
-    });
-
-    static PACKAGE_METADATA_ID: &str = "metadata-helper 0.1.0 (path+file:///Users/fakeuser/local/testcrates/metadata/metadata-helper)";
-    fn package_metadata() -> PackageMetadata<'static> {
-        PACKAGE_GRAPH_FIXTURE
-            .metadata(&PackageId::new(PACKAGE_METADATA_ID))
-            .expect("package ID is valid")
     }
 
     #[test]
